- Migrate to cocos2d-x 3.0 STABLE

+++ cocos2d/cocos/2d/CCTMXLayer.cpp
@@ -34,6 +34,8 @@ THE SOFTWARE.
 #include "ccCArray.h"
 #include "CCDirector.h"
 
+#include "deprecated/CCString.h" // For StringUtils::format
+
 NS_CC_BEGIN
 
 
@@ -492,7 +494,7 @@ void TMXLayer::setTileGID(uint32_t gid, const Point& pos, TMXTileFlags flags)
 {
     CCASSERT(pos.x < _layerSize.width && pos.y < _layerSize.height && pos.x >=0 && pos.y >=0, "TMXLayer: invalid position");
     CCASSERT(_tiles && _atlasIndexArray, "TMXLayer: the tiles map has been released");
-    CCASSERT(gid == 0 || gid >= _tileSet->_firstGid, "TMXLayer: invalid gid" );
+    CCASSERT(gid == 0 || (int)gid >= _tileSet->_firstGid, "TMXLayer: invalid gid" );
 
     TMXTileFlags currentFlags;
     uint32_t currentGID = getTileGIDAt(pos, &currentFlags);
